backend/requirements.txt
amqp
aniso8601
async-timeout
Babel
billiard
blinker
Brotli
cachelib
celery
cffi
click
click-didyoumean
click-plugins
click-repl
cssselect2
dnspython
email-validator
Flask
Flask-BabelEx
Flask-Caching
Flask-Login
Flask-Mail
Flask-Principal
Flask-RESTful
Flask-Security-Too
Flask-SQLAlchemy
Flask-WTF
fonttools
greenlet
html5lib
idna
importlib-resources
itsdangerous
Jinja2
kombu
MarkupSafe
passlib
Pillow
prompt-toolkit
pycparser
pydyf
pyphen
python-dateutil
pytz
redis
six
speaklater
SQLAlchemy
tinycss2
typing_extensions
tzdata
vine
wcwidth
weasyprint
webencodings
Werkzeug
WTForms
zopfli
httplib2
sckit-learn
pandas
MailHog